+++ src/pyduck_ona_profile/__init__.py
@@ -48,7 +48,7 @@ from pyduck_ona_profile.schema import (
 from pyduck_ona_profile.subject import Profile, Subject
 from pyduck_ona_profile.timeline import Timeline
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"
 
 __all__ = [
     "AskResult",